# perf: improve PQ computing distances