--- a/rust/lance-index/Cargo.toml
+++ b/rust/lance-index/Cargo.toml
@@ -83,6 +83,10 @@ name = "pq_dist_table"
 harness = false
 
 [[bench]]
+name = "4bitpq_dist_table"
+harness = false
+
+[[bench]]
 name = "pq_assignment"
 harness = false