@chebbyChefNEQ 2024-11-21
another qq: would SQ benefit from the same optimization? Let's try it?