@BubbleCal 2024-11-22
> another qq: would SQ benefit from the same optimization? Let's try it?

no, distance computing for SQ is not with the same problem